Add Qt and CMake
- Moved source code to src dir
- Added CMake
- Removed Makefile
- Changed MySQL database to QtDatabase (SQLite)
- Code formate
