+++ .gitignore
@@ -1,3 +1,5 @@
+build
 .vs
 x64
 sf_chat.*
+*.user
