+++ src/message.cpp
@@ -3,13 +3,13 @@
 #include <iostream>
 
 const std::string& Message::GetFrom() const {
-	return m_from;
+  return m_from;
 }
 
 const std::string& Message::GetTo() const {
-	return m_to;
+  return m_to;
 }
 
 const std::string& Message::GetText() const {
-	return m_text;
+  return m_text;
 }
